beam_app/requirements.txt
pyserial==3.5
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dotenv==1.2.2
shiboken6==6.11.0
